Update squba packages to align with most recent GitHub commits

Update squba packages to align with most recent GitHub commits

Usage

squba_update()

Value

a message in the console with information about which packages need to be updated based on what is currently installed

Examples

if (FALSE) { # \dontrun{
## Requires that a GitHub PAT is established

squba_update()
} # }